rename package and wire resource
