+++ service/controller/resource/release/create.go
@@ -1,4 +1,4 @@
-package accountid
+package release
 
 import (
 	"context"
